fix: move file to components so next doesn't consider it as page
